linkedin_outreach/requirements.txt
playwright>=1.47.0
playwright-stealth>=1.0.6
python-dotenv>=1.0.1
apscheduler>=3.10.4
